Add files via upload
In this project, I collected reviews of gyms located in Jersey City, NJ, using the Yelp API. I utilized TextBlob and the Naive Bayes Analyzer to classify reviews as positive, negative, or neutral. Additionally, I identified the top five gyms with the highest ratings based on the number of positive sentiment reviews.